fix: sorting reset after filtering
